Split input classes (#2577)
